refactor: 패키지 이동 및 코드 리팩터링
