+++ src/test/java/com/kau/capstone/domain/auth/access/TestAccessToken.java
@@ -1,9 +1,9 @@
-package com.kau.capstone.common.logout;
+package com.kau.capstone.domain.auth.access;
 
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
 
 @Component
 @Profile("test")
-public class TestLogout {
+public class TestAccessToken {
 }
