+++ src/test/java/com/kau/capstone/domain/auth/logout/TestLogout.java
@@ -1,9 +1,9 @@
-package com.kau.capstone.common.refresh;
+package com.kau.capstone.domain.auth.logout;
 
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
 
 @Component
 @Profile("test")
-public class TestRefreshToken {
+public class TestLogout {
 }
