+++ src/test/java/com/kau/capstone/domain/auth/refresh/TestRefreshToken.java
@@ -1,9 +1,9 @@
-package com.kau.capstone.common.redirect;
+package com.kau.capstone.domain.auth.refresh;
 
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
 
 @Component
 @Profile("test")
-public class TestRedirect {
+public class TestRefreshToken {
 }
