+++ src/test/java/com/kau/capstone/global/common/ControllerTest.java
@@ -1,0 +1,25 @@
+package com.kau.capstone.global.common;
+
+import com.kau.capstone.domain.member.repository.MemberRepository;
+import io.restassured.RestAssured;
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.web.server.LocalServerPort;
+import org.springframework.test.annotation.DirtiesContext;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_EACH_TEST_METHOD)
+public class ControllerTest {
+
+    @LocalServerPort
+    private int port;
+
+    @Autowired
+    protected MemberRepository memberRepository;
+
+    @BeforeEach
+    void setUp() {
+        RestAssured.port = port;
+    }
+}
